# api/src/openapi.yaml
openapi: 3.0.3
info:
  title: CTD API (OSS)
  version: 1.0.0
  description: |
    Crash to Desktop Reporter - Self-hosted API for collecting and analyzing game crash reports.

    ## Authentication
    Most endpoints require an API key passed via the `x-api-key` header.
    Generate keys using the `/api-keys` endpoint.
  contact:
    name: CTD Support
    url: https://ctd.ezmode.games
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Health
    description: Server health checks
  - name: API Keys
    description: API key management
  - name: Config
    description: Configuration file download
  - name: Crashes
    description: Crash report submission and retrieval

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Check if the server is running
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: 0.1.0

  /api-keys:
    post:
      tags: [API Keys]
      summary: Generate new API key
      description: |
        Create a new API key. The full key is only returned once on creation.
        Store it securely - it cannot be retrieved later.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Friendly name for the key
                  default: default
                  maxLength: 100
                expiresIn:
                  type: integer
                  description: Expiration time in milliseconds from now
                  example: 86400000
            example:
              name: production
              expiresIn: 31536000000
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [API Keys]
      summary: List API keys
      description: List all API keys. Only the key prefix is shown for security.
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKeyInfo'

  /api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke API key
      description: Permanently delete an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID (ULID)
      responses:
        '200':
          description: Key revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/NotFound'

  /config:
    get:
      tags: [Config]
      summary: Download config file
      description: |
        Download a pre-configured ctd.toml file with your API key embedded.
        This file should be placed in your game's directory.
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
          description: Your API key
          example: ctd_a8Kj2mNp4qRs6tUv8wXy0zB3dF5gH7jL
        - name: format
          in: query
          schema:
            type: string
            enum: [toml, json]
            default: toml
          description: Output format
      responses:
        '200':
          description: Configuration file
          content:
            application/toml:
              schema:
                type: string
                example: |
                  [api]
                  url = "https://your-server.example.com"
                  api_key = "ctd_a8Kj2mNp4qRs6tUv8wXy0zB3dF5gH7jL"
            application/json:
              schema:
                type: object
                properties:
                  server_url:
                    type: string
                  api_key:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /crashes:
    post:
      tags: [Crashes]
      summary: Submit crash report
      description: Submit a new crash report from a game client
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCrashReport'
      responses:
        '201':
          description: Crash report submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrashReportCreated'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /crashes/{id}:
    get:
      tags: [Crashes]
      summary: Get crash report
      description: Retrieve a crash report by ID. Requires share token for non-public reports.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Crash report ID (ULID)
        - name: token
          in: query
          schema:
            type: string
          description: Share token for accessing non-public reports
      responses:
        '200':
          description: Crash report details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrashReport'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication

  schemas:
    ApiKeyCreated:
      type: object
      properties:
        key:
          type: string
          description: The full API key (only shown once)
          example: ctd_a8Kj2mNp4qRs6tUv8wXy0zB3dF5gH7jL
        id:
          type: string
          description: Key ID for management
        name:
          type: string
        keyPrefix:
          type: string
          example: ctd_a8Kj
        expiresAt:
          type: integer
          nullable: true
          description: Expiration timestamp (ms since epoch)
        createdAt:
          type: integer

    ApiKeyInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        keyPrefix:
          type: string
          example: ctd_a8Kj
        lastUsedAt:
          type: integer
          nullable: true
        expiresAt:
          type: integer
          nullable: true
        createdAt:
          type: integer

    CreateCrashReport:
      type: object
      required:
        - gameId
        - stackTrace
        - gameVersion
        - loadOrderJson
        - pluginCount
        - crashedAt
      properties:
        schemaVersion:
          type: integer
          default: 1
        gameId:
          type: string
          description: Game identifier (e.g., skyrim, fallout4)
          example: skyrim
        stackTrace:
          type: string
          description: Stack trace from the crash
          maxLength: 100000
        crashHash:
          type: string
          description: Pre-computed crash hash (optional)
          maxLength: 64
        exceptionCode:
          type: string
          example: "0xC0000005"
        exceptionAddress:
          type: string
        faultingModule:
          type: string
          example: SkyrimSE.exe
        gameVersion:
          type: string
          example: 1.6.1170
        scriptExtenderVersion:
          type: string
          example: 2.2.3
        osVersion:
          type: string
          example: Windows 10.0.19045
        loadOrderJson:
          type: string
          description: JSON array of load order
          example: '[{"name":"Skyrim.esm","enabled":true,"index":0}]'
        pluginCount:
          type: integer
          minimum: 0
        crashedAt:
          type: integer
          description: Timestamp when crash occurred (ms since epoch)
        notes:
          type: string
          maxLength: 5000

    CrashReportCreated:
      type: object
      properties:
        id:
          type: string
          description: Crash report ID
        shareToken:
          type: string
          description: Token for sharing the report

    CrashReport:
      type: object
      properties:
        id:
          type: string
        schemaVersion:
          type: integer
        gameId:
          type: string
        crashHash:
          type: string
        stackTrace:
          type: string
        exceptionCode:
          type: string
        exceptionAddress:
          type: string
        faultingModule:
          type: string
        gameVersion:
          type: string
        scriptExtenderVersion:
          type: string
        osVersion:
          type: string
        loadOrder:
          type: array
          items:
            type: object
        pluginCount:
          type: integer
        crashedAt:
          type: integer
        submittedAt:
          type: integer
        isPublic:
          type: boolean
        notes:
          type: string
        pattern:
          type: object
          nullable: true
          properties:
            id:
              type: string
            patternName:
              type: string
            occurrenceCount:
              type: integer
            knownFix:
              type: string

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            issues:
              type: array
              items:
                type: object

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Missing required parameter

    Unauthorized:
      description: Invalid or expired API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid API key

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found